build: split runtime tool planning crate
