+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/hermes-core",
     "crates/hermes-provider-runtime",
     "crates/hermes-app-runtime",
+    "crates/hermes-tool-planning",
     "crates/hermes-agent",
     "crates/hermes-tools",
     "crates/hermes-gateway",
@@ -36,6 +37,7 @@ repository = "https://github.com/sheawinkler/hermes-agent-ultra"
 hermes-core = { path = "crates/hermes-core" }
 hermes-provider-runtime = { path = "crates/hermes-provider-runtime" }
 hermes-app-runtime = { path = "crates/hermes-app-runtime" }
+hermes-tool-planning = { path = "crates/hermes-tool-planning" }
 hermes-agent = { path = "crates/hermes-agent" }
 hermes-tools = { path = "crates/hermes-tools" }
 hermes-gateway = { path = "crates/hermes-gateway" }
