+++ crates/hermes-cli/src/app.rs
@@ -2100,7 +2100,7 @@ impl App {
         wire_cron_scheduler_backend(&tool_registry, cron_scheduler.clone());
         let agent_tool_registry = Arc::new(bridge_tool_registry(&tool_registry));
         let tool_schemas =
-            crate::platform_toolsets::resolve_platform_tool_schemas(&config, "cli", &tool_registry);
+            hermes_tool_planning::resolve_platform_tool_schemas(&config, "cli", &tool_registry);
 
         let session_id = Uuid::new_v4().to_string();
         let mut agent_config = build_agent_config(&config, &current_model);
