+++ crates/hermes-tool-planning/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "hermes-tool-planning"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Runtime platform tool-planning policy for Hermes Agent"
+
+[dependencies]
+hermes-agent = { workspace = true }
+hermes-config = { workspace = true }
+hermes-core = { workspace = true }
+hermes-tools = { workspace = true }
+serde_json = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+async-trait = { workspace = true }
+tempfile = "3"
