+++ docs/parity/behavioral-similarity-diff.md
@@ -1,6 +1,6 @@
 # Behavioral Similarity Diff
 
-Generated: `2026-06-23T10:14:45.420946+00:00`
+Generated: `2026-06-23T11:00:13.601707+00:00`
 
 ## Gate
