+++ docs/parity/deep-problem-solving-diff.md
@@ -1,6 +1,6 @@
 # Deep Problem-Solving Diff
 
-Generated: `2026-06-23T10:14:47.827059+00:00`
+Generated: `2026-06-23T11:00:16.127584+00:00`
 
 ## Gate
