+++ docs/parity/global-parity-proof.json
@@ -260,7 +260,7 @@
       "unverified_cases": 0
     }
   },
-  "generated_at_utc": "2026-06-23T10:14:49.130616+00:00",
+  "generated_at_utc": "2026-06-23T11:00:17.533132+00:00",
   "gpar_completion": {
     "GPAR-01": true,
     "GPAR-02": true,
@@ -538,7 +538,7 @@
       "queue_pending": 0,
       "queue_total": 6132,
       "rust_test_files": 314,
-      "rust_test_functions": 3629,
+      "rust_test_functions": 3630,
       "test_intent_mapping_ratio": 1.0,
       "test_intents_mapped": 10,
       "test_intents_total": 10,
