+++ docs/parity/test-intent-mapping.md
@@ -1,6 +1,6 @@
 # Test Intent Mapping
 
-Generated: `2026-06-23T10:14:42.605295+00:00`
+Generated: `2026-06-23T10:59:20.156630+00:00`
 
 | Intent | Mapped | Evidence Count |
 | --- | --- | ---: |
