+++ docs/parity/workstream-status.json
@@ -1,6 +1,6 @@
 {
-  "generated_at_utc": "2026-06-23T03:25:28-06:00",
-  "head_sha": "b035fbba85aadb59051ed9172f65e6375e06d654",
+  "generated_at_utc": "2026-06-23T04:15:52-06:00",
+  "head_sha": "3920d29c8217942c688847e760479f2af97c019d",
   "states": {
     "complete": 7
   },
