+++ docs/parity/workstream-status.md
@@ -1,6 +1,6 @@
 # Workstream Status
 
-- Local HEAD: `b035fbba85aadb59051ed9172f65e6375e06d654`
+- Local HEAD: `3920d29c8217942c688847e760479f2af97c019d`
 - Upstream: `upstream/main` (`5ecf3bf0e0726b8b33682bb5c3aad9679b7b5be4`)
 
 | Workstream | Title | State |
